feat: [features in development] add get device by id, get devices and get history and save user data  (#40)
* feat: add get device by id, get devices and get history - GraphQL (#33)

* Added the types used by devices, the getDEvices query and resolvers

* Adding getDeviceById query and resolver

* getDevice and getDeviceById with tests

* Adding getDeviceHistory to Resolvers

* getDeviceHistory returning array of History

* Added user configuration files and reference to Schema

* added suggested configuration in package.json and device resolver

* added suggested configuration in package.json and device resolver

* Update package.json

* removed history queries

* removed history queries

* removed history queries

* adjusted queries to right typedef version

* Reimplemented getDeviceHistory

* Unit tests for device resolvers

* small fixes

* Added new unit tests, comments on types and removed lines from Schema

* Updated Schema, Resolvers and TypeDefs, and added Unit tests

* Added/adjusted unit tests

* somefix

* fixed tests and device resolvers

* fixes


* added user directory

* Corrected query strings and error treatment

* Adding tests and more resolver corrections

* Added check before running in dev and error throw if null in Resolvers

* upgrade codecov version

* Updated databaseCheck and db module references in different files.

* Corrections

Added backstage exclusive env variables and config, hardcoded 'user_config' database name and updated jest call in package.json

* Updated Readme and documentation

Added environment variables table in readme, updated Schema and user types documentation.

* Changed environment variables' names to more descriptive ones

* change route

* Added generic user option to resolver

* User not obligatory

* add some check


* fix problems

* fix jest

+++ src/travis/deploy-gh-pages.sh
@@ -1,25 +1,37 @@
 #!/usr/bin/env bash
-if [ "${TRAVIS_BRANCH}" == "master" ]
-then
-  export VERSION=""
-else
-  export VERSION="${TRAVIS_BRANCH}"
+if [ $TRAVIS_PULL_REQUEST == false ] ; then
+ 
+  if [ "${TRAVIS_BRANCH}" == "master" ]
+  then
+    export VERSION=""
+  else
+    export VERSION="${TRAVIS_BRANCH}"
+  fi
+
+  git remote add -t gh-pages -f origin-gh-pages https://github.com/${TRAVIS_REPO_SLUG}
+  git fetch origin-gh-pages
+  git checkout .
+  git checkout gh-pages
+  
+  git checkout ${TRAVIS_BRANCH} -- ./src
+
+  node ./src/docs/GenerateFileSchema.js
+  graphdoc -s ./src/docs/schemaDoc.graphql -o ./${VERSION} -f
+  
+  rm -fr ./_static
+  git rm -r ./src -f
+
+  git add ${VERSION}
+  git status
+  ls
+  git commit -m 'Updating gh-pages' 
+   git status
+  ls
+  git push --force http://${GITHUB_TOKEN}:x-oauth-basic@github.com/${TRAVIS_REPO_SLUG} gh-pages
+  
+  
+  git checkout -- .
+  git clean -fd
+  git checkout ${TRAVIS_BRANCH}
+  
 fi
-
-git remote add -t gh-pages -f origin-gh-pages https://github.com/${TRAVIS_REPO_SLUG}
-git fetch origin-gh-pages
-git checkout -- .
-git checkout gh-pages
-git checkout ${TRAVIS_BRANCH} -- ./src
-
-node ./src/docs/GenerateFileSchema.js
-graphdoc -s ./src/docs/schemaDoc.graphql -o ./${VERSION} -f
-rm -fr ./_static
-git rm -r src -f
-
-git add ./${VERSION}
-git commit -m 'Updating gh-pages' --amend
-git push --force http://${GITHUB_TOKEN}:x-oauth-basic@github.com/${TRAVIS_REPO_SLUG} gh-pages
-git checkout -- .
-git clean -fd
-git checkout ${TRAVIS_BRANCH}
